+++ src/main.ts
@@ -1,3 +1,4 @@
-import { Server } from "./server/server.ts";
+import { Server } from "./server.ts";
+import { Lsp } from "./lsp/lsp.ts";
 
-Server.serve();
+Server.serve(Lsp);
